// src/api/schemas/mediaSchema.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Keyo API
 * API for Keyo Polls app
 * OpenAPI spec version: 1.0.0
 */
import type { MediaSchemaAltText } from './mediaSchemaAltText';
import type { MediaSchemaCreatedAt } from './mediaSchemaCreatedAt';
import type { MediaSchemaDuration } from './mediaSchemaDuration';
import type { MediaSchemaFileUrl } from './mediaSchemaFileUrl';
import type { MediaSchemaHeight } from './mediaSchemaHeight';
import type { MediaSchemaOrder } from './mediaSchemaOrder';
import type { MediaSchemaThumbnailUrl } from './mediaSchemaThumbnailUrl';
import type { MediaSchemaWidth } from './mediaSchemaWidth';

/**
 * Schema for media attachments
 */
export interface MediaSchema {
  id: number;
  media_type: string;
  file_url?: MediaSchemaFileUrl;
  thumbnail_url?: MediaSchemaThumbnailUrl;
  width?: MediaSchemaWidth;
  height?: MediaSchemaHeight;
  alt_text?: MediaSchemaAltText;
  duration?: MediaSchemaDuration;
  order?: MediaSchemaOrder;
  created_at?: MediaSchemaCreatedAt;
}